requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.26.0
Django==3.2.6
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-createsuperuserwithpassword==2.0.0
django-filter==2.4.0
django-js-asset==1.2.2
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
